+++ InventoryApi/InventoryApi/Controllers/UserController.cs
@@ -1,4 +1,5 @@
-﻿using InventoryApi.Entities;
+﻿using System.Security.Claims;
+using InventoryApi.Entities;
 using InventoryApi.Models.UserDtos;
 using InventoryApi.Services.UserServices;
 using Microsoft.AspNetCore.Authorization;
@@ -38,6 +39,10 @@ namespace InventoryApi.Controllers
         [HttpDelete("delete/{userId}")]
         public async Task<ActionResult<bool>> DeleteAccount(int userId)
         {
+            var tokenUserId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)!.Value);
+            if (tokenUserId != userId) return Forbid();
+
+
             bool userDeleted = await userService.DeleteUserAsync(userId);
 
             if (!userDeleted) return BadRequest("User not found");
@@ -51,6 +56,10 @@ namespace InventoryApi.Controllers
         [HttpPut("update-basic/{userId}")]
         public async Task<ActionResult<bool>> UpdateBasicInformation(int userId, SetUserInfoDto request)
         {
+            var tokenUserId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)!.Value);
+            if (tokenUserId != userId) return Forbid();
+
+
             bool basicInfoUpdated = await userService.UpdateBasicInformationAsync(userId, request);
             if (!basicInfoUpdated) return BadRequest("User not found");
 
@@ -63,6 +72,10 @@ namespace InventoryApi.Controllers
         [HttpPut("update-username/{userId}")]
         public async Task<ActionResult<bool>> UpdateUsername(int userId, UsernameDto request)
         {
+            var tokenUserId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)!.Value);
+            if (tokenUserId != userId) return Forbid();
+
+
             bool usernameUpdated = await userService.UpdateUsernameAsync(userId, request);
             if (!usernameUpdated) return BadRequest("User not found");
 
@@ -75,6 +88,10 @@ namespace InventoryApi.Controllers
         [HttpPut("update-email/{userId}")]
         public async Task<ActionResult<bool>> UpdateEmail(int userId, EmailDto request)
         {
+            var tokenUserId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)!.Value);
+            if (tokenUserId != userId) return Forbid();
+
+
             bool emailUpdated = await userService.UpdateEmailAsync(userId,request);
             if (!emailUpdated) return BadRequest("Email not updated");
 
@@ -87,6 +104,10 @@ namespace InventoryApi.Controllers
         [HttpPut("update-password/{userId}")]
         public async Task<ActionResult<bool>> UpdatePassword(int userId, PasswordDto request)
         {
+            var tokenUserId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)!.Value);
+            if (tokenUserId != userId) return Forbid();
+
+
             bool passowordUpdated = await userService.UpdatePasswordAsync(userId, request);
             if (!passowordUpdated) return BadRequest("Password not updated");
 
@@ -99,6 +120,10 @@ namespace InventoryApi.Controllers
         [HttpGet("get-user/{userId}")]
         public async Task<ActionResult<UserDto>> GetUser(int userId)
         {
+            var tokenUserId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)!.Value);
+            if (tokenUserId != userId) return Forbid();
+
+
             GetUserDto? user = await userService.GetUserAsync(userId);
             if (user == null) return BadRequest("User not found");
